+++ src/Instance/HoldingsList/Holding/HoldingButtonsGroup/index.js
@@ -1,0 +1,4 @@
+export { default as AddItemButton } from './AddItemButton';
+export { default as HoldingButtonsGroup } from './HoldingButtonsGroup';
+export { default as ItemsCountBadge } from './ItemsCountBadge';
+export { default as ViewHoldingsButton } from './ViewHoldingsButton';