+++ src/components/InstancesList/InstancesList.js
@@ -724,17 +724,15 @@ class InstancesList extends React.Component {
   setSegmentSortBy = (sortBy) => {
     const { segment } = this.props;
 
-    const segmentsSortBy = this.state.segmentsSortBy.map((key) => {
-      if (key.name === segment) {
-        key.sort = sortBy;
+    this.setState(prevState => ({
+      segmentsSortBy: prevState.segmentsSortBy.map((key) => {
+        if (key.name === segment) {
+          key.sort = sortBy;
+          return key;
+        }
         return key;
-      }
-      return key;
-    });
-
-    this.setState({
-      segmentsSortBy
-    });
+      }),
+    }));
   }
 
   getActionMenu = ({ onToggle }) => {
@@ -751,6 +749,7 @@ class InstancesList extends React.Component {
     const visibleColumns = this.getVisibleColumns();
     const columnMapping = this.getColumnMapping();
     const canExportMarc = stripes.hasPerm('ui-data-export.app.enabled');
+    const canCreateItemsInTransitReport = stripes.hasPerm('ui-inventory.items.create-in-transit-report');
 
     const buildOnClickHandler = onClickHandler => {
       return () => {
@@ -841,7 +840,7 @@ class InstancesList extends React.Component {
               icon: 'report',
               messageId: 'ui-inventory.exportInProgress',
             }) :
-            !checkIfUserInCentralTenant(stripes) && this.getActionItem({
+            canCreateItemsInTransitReport && !checkIfUserInCentralTenant(stripes) && this.getActionItem({
               id: 'dropdown-clickable-get-report',
               icon: 'report',
               messageId: 'ui-inventory.inTransitReport',