+++ src/hooks/index.js
@@ -12,3 +12,4 @@ export { default as useLastSearchTerms } from './useLastSearchTerms';
 export { default as useSearchForShadowInstanceTenants } from './useSearchForShadowInstanceTenants';
 export { default as useUserTenantPermissions } from './useUserTenantPermissions';
 export { default as useLocationsForTenants } from './useLocationsForTenants';
+export { default as useClassificationIdentifierTypes } from './useClassificationIdentifierTypes';