+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Remove unused code related to auto-open record detail view. Refs UIIN-2819.
 * Replace `all` with the `=` operator to get correct results when using the `All` search option.
+* Create new settings for classification type sorting. Refs UIIN-2775.
 
 ## [10.0.1] (IN PROGRESS)