+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Do not remove field from the form when its value is an empty string. Fixes UIIN-2787.
 * ECS - Member consortial accordion is not displaying when user have according affiliation but don't have permissions for view holdings. Fixes UIIN-2780.
 * Do not remove field from the form when its value is an empty string. Fixes UIIN-2787.
+* ECS: Shared instance cannot be edited from member tenant, even with permissions in both Central and member tenants. Fixes UIIN-2832.
 
 ## [11.0.0](https://github.com/folio-org/ui-inventory/tree/v11.0.0) (2024-03-21)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v10.0.11...v11.0.0)